requirements.txt
Django==2.1.7
django-tinymce==3.2.0
djangorestframework==3.8.2
joblib==1.0.1
mysqlclient==2.0.3
numpy==1.20.1
pandas==1.2.2
Pillow==8.1.0
PyMySQL==1.0.2
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
